+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_dynamic/visitor.dart
@@ -10,13 +10,17 @@ class _Visitor extends RecursiveAstVisitor<void> {
     final parent = node.parent;
     if (parent is NamedType && (parent.type?.isDynamic ?? false)) {
       final grandParent = node.parent?.parent;
-      if (grandParent != null) {
-        final grandGrandParent = grandParent.parent;
-        if (!(grandGrandParent is NamedType &&
-            (grandGrandParent.type?.isDartCoreMap ?? false))) {
-          _nodes.add(grandParent);
-        }
+      if (grandParent != null && !_isWithinMap(grandParent)) {
+        _nodes.add(grandParent);
       }
     }
+  }
+
+  bool _isWithinMap(AstNode grandParent) {
+    final grandGrandParent = grandParent.parent;
+
+    return grandGrandParent is NamedType &&
+            (grandGrandParent.type?.isDartCoreMap ?? false) ||
+        grandGrandParent is SetOrMapLiteral && grandGrandParent.isMap;
   }
 }