+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_dynamic/examples/example.dart
@@ -35,4 +35,9 @@ class SomeClass {
 
 abstract class BaseClass<T> {}
 
-class Generic extends BaseClass<dynamic> {} // LINT
+// LINT
+class Generic extends BaseClass<dynamic> {
+  final Map<String, dynamic> myMap = {};
+
+  final myMap = <String, dynamic>{};
+}