+++ lib/src/analyzers/lint_analyzer/anti_patterns/anti_patterns_list/long_method.dart
@@ -34,7 +34,7 @@ class LongMethod extends Pattern {
         ),
         super(
           id: patternId,
-          severity: readSeverity(patternSettings, Severity.none),
+          severity: readSeverity(patternSettings, Severity.warning),
           excludes: readExcludes(patternSettings),
         );